# sml_shared/src/main/java/com/loblaw/metrics/shared/model/LoadAverage.java
package com.loblaw.metrics.shared.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@JsonIgnoreProperties(ignoreUnknown = true)
public class LoadAverage {
	@JsonProperty("1minute")
	Double oneMinute;
	@JsonProperty("5minutes")
	Double fiveMinutes;
	@JsonProperty("15minutes")
	Double fifteenMinutes;
}